Update alloverlay.py
Apart of fixing some minor errors and implementating the image undistortion system, i have greatly improved the efficiency of the algorythm, as now the data get's streamed through just one "cali" and one "corn" object instead of two of every type. This happens because both cameras are identical, so different calibration or processign parameters are completely unnecessary, and in fact a possible issue. I remain uncertain about the way of implementing the feature matcher, although it would work more coherently with this architecture because of the fact that it would require streaming both images into the same object.